requirements.txt
asgiref==3.8.1
astroid==3.3.5
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
datetime-truncate==1.1.1
dill==0.3.9
distlib==0.3.9
Django==5.1.3
django-appconf==1.0.6
django-bower==5.2.0
django-htmx==1.21.0
django-imagekit==5.0.0
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-unfold==0.41.0
exceptiongroup==1.2.2
filelock==3.16.1
greenlet==3.1.1
identify==2.6.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pilkit==3.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
pre_commit==4.0.1
pyee==12.0.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-base-url==2.1.0
pytest-django==4.9.0
pytest-playwright==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
